+++ httsleep/main.py
@@ -12,6 +12,7 @@ from ._compat import string_types
 DEFAULT_POLLING_INTERVAL = 2 # in seconds
 DEFAULT_MAX_RETRIES = 50
 VALID_CONDITIONS = ['status_code', 'json', 'jsonpath', 'text', 'callback']
+DEFAULT_SESSION = requests.Session()
 
 
 class HttSleeper(object):
@@ -33,10 +34,14 @@ class HttSleeper(object):
     :param callback: shorthand for a success condition dependent on a callback
                      function that takes the response as an argument returning True.
     :param auth: a (username, password) tuple for HTTP authentication.
-    :param headers: a dict of HTTP headers.
+    :param headers: a dict of HTTP headers.  If specified, these will be merged with (and take
+                    precedence over) any headers provided in the session.
+    :param session: a Requests session, providing cookie persistence, connection-pooling, and
+                    configuration (e.g. headers).
     :param verify: Either a boolean, in which case it controls whether we verify the server's
                    TLS certificate, or a string, in which case it must be a path to a CA
-                   bundle to use. Defaults to ``True``.
+                   bundle to use. If specified, this takes precedence over any value defined
+                   in the session (which itself would be ``True``, by default).
     :param polling_interval: how many seconds to sleep between requests.
     :param max_retries: the maximum number of retries to make, after which
                         a StopIteration exception is raised.
@@ -49,7 +54,7 @@ class HttSleeper(object):
     """
     def __init__(self, url_or_request, until=None, alarms=None,
                  status_code=None, json=None, jsonpath=None, text=None, callback=None,
-                 auth=None, headers=None, verify=True,
+                 auth=None, headers=None, session=DEFAULT_SESSION, verify=None,
                  polling_interval=DEFAULT_POLLING_INTERVAL,
                  max_retries=DEFAULT_MAX_RETRIES,
                  ignore_exceptions=None,
@@ -87,11 +92,13 @@ class HttSleeper(object):
                          'jsonpath': jsonpath, 'text': text, 'callback': callback}
             until.append({k: v for k, v in condition.items() if v})
 
-        self.verify = verify
+        self.kwargs = {}
+        if verify is not None:
+            self.kwargs['verify'] = verify
         self.until = until
         self.alarms = alarms
         self.polling_interval = int(polling_interval)
-        self.session = requests.Session()
+        self.session = session
         self.log = logging.getLogger()
         self.log.setLevel(loglevel)
 
@@ -149,7 +156,7 @@ class HttSleeper(object):
         """
         while True:
             try:
-                response = self.session.send(self.request.prepare(), verify=self.verify)
+                response = self.session.send(self.session.prepare_request(self.request), **self.kwargs)
                 for condition in self.alarms:
                     if self.meets_condition(response, condition):
                         raise Alarm(response, condition)
@@ -198,7 +205,7 @@ class HttSleeper(object):
 
 def httsleep(url_or_request, until=None, alarms=None, status_code=None,
              json=None, jsonpath=None, text=None, callback=None,
-             auth=None, headers=None, verify=True,
+             auth=None, headers=None, session=DEFAULT_SESSION, verify=None,
              polling_interval=DEFAULT_POLLING_INTERVAL,
              max_retries=DEFAULT_MAX_RETRIES,
              ignore_exceptions=None,
@@ -211,7 +218,7 @@ def httsleep(url_or_request, until=None, alarms=None, status_code=None,
     return HttSleeper(
         url_or_request, until=until, alarms=alarms, status_code=status_code,
         json=json, jsonpath=jsonpath, text=text, callback=callback,
-        auth=auth, headers=headers, verify=verify,
+        auth=auth, headers=headers, session=session, verify=verify,
         polling_interval=polling_interval,
         max_retries=max_retries,
         ignore_exceptions=ignore_exceptions,